Jit64: Move X64 stuff out of JitCommon.
