+++ Source/Core/Core/Core.vcxproj.filters
@@ -627,8 +627,8 @@
     <ClCompile Include="PowerPC\SignatureDB.cpp">
       <Filter>PowerPC</Filter>
     </ClCompile>
-    <ClCompile Include="PowerPC\JitCommon\Jit_Util.cpp">
-      <Filter>PowerPC\JitCommon</Filter>
+    <ClCompile Include="PowerPC\Jit64Common\Jit_Util.cpp">
+      <Filter>PowerPC\Jit64Common</Filter>
     </ClCompile>
     <ClCompile Include="PowerPC\JitCommon\JitAsmCommon.cpp">
       <Filter>PowerPC\JitCommon</Filter>
@@ -639,8 +639,8 @@
     <ClCompile Include="PowerPC\JitCommon\JitCache.cpp">
       <Filter>PowerPC\JitCommon</Filter>
     </ClCompile>
-    <ClCompile Include="PowerPC\JitCommon\TrampolineCache.cpp">
-      <Filter>PowerPC\JitCommon</Filter>
+    <ClCompile Include="PowerPC\Jit64Common\TrampolineCache.cpp">
+      <Filter>PowerPC\Jit64Common</Filter>
     </ClCompile>
     <ClCompile Include="PowerPC\Jit64IL\IR_X86.cpp">
       <Filter>PowerPC\JitIL</Filter>
@@ -717,8 +717,8 @@
     <ClCompile Include="PowerPC\MMU.cpp">
       <Filter>PowerPC</Filter>
     </ClCompile>
-    <ClCompile Include="PowerPC\JitCommon\JitBackpatch.cpp">
-      <Filter>PowerPC\JitCommon</Filter>
+    <ClCompile Include="PowerPC\Jit64Common\JitBackpatch.cpp">
+      <Filter>PowerPC\Jit64Common</Filter>
     </ClCompile>
     <ClCompile Include="PowerPC\Jit64Common\Jit64AsmCommon.cpp">
       <Filter>PowerPC\Jit64Common</Filter>
@@ -1199,8 +1199,8 @@
     <ClInclude Include="PowerPC\SignatureDB.h">
       <Filter>PowerPC</Filter>
     </ClInclude>
-    <ClInclude Include="PowerPC\JitCommon\Jit_Util.h">
-      <Filter>PowerPC\JitCommon</Filter>
+    <ClInclude Include="PowerPC\Jit64Common\Jit_Util.h">
+      <Filter>PowerPC\Jit64Common</Filter>
     </ClInclude>
     <ClInclude Include="PowerPC\JitCommon\JitAsmCommon.h">
       <Filter>PowerPC\JitCommon</Filter>
@@ -1211,8 +1211,8 @@
     <ClInclude Include="PowerPC\JitCommon\JitCache.h">
       <Filter>PowerPC\JitCommon</Filter>
     </ClInclude>
-    <ClInclude Include="PowerPC\JitCommon\TrampolineCache.h">
-      <Filter>PowerPC\JitCommon</Filter>
+    <ClInclude Include="PowerPC\Jit64Common\TrampolineCache.h">
+      <Filter>PowerPC\Jit64Common</Filter>
     </ClInclude>
     <ClInclude Include="PowerPC\Jit64IL\JitIL.h">
       <Filter>PowerPC\JitIL</Filter>
