+++ Source/Core/Core/PowerPC/Jit64/Jit_LoadStoreFloating.cpp
@@ -9,7 +9,7 @@
 #include "Common/x64Emitter.h"
 #include "Core/ConfigManager.h"
 #include "Core/PowerPC/Jit64/JitRegCache.h"
-#include "Core/PowerPC/JitCommon/Jit_Util.h"
+#include "Core/PowerPC/Jit64Common/Jit_Util.h"
 
 using namespace Gen;
